+++ vimgolf/commands/local.py
@@ -2,7 +2,7 @@ import os
 import tempfile
 
 from vimgolf import logger
-from vimgolf.core import Challenge
+from vimgolf.challenge import Challenge
 from vimgolf.play import play
 from vimgolf.utils import format_
